apps/requirements.txt
Django>=4.2.11
requests>=2.31.0
beautifulsoup4>=4.12.3
jieba>=0.42.1
jieba3k>=0.35.1
diskcache>=5.6.3
PyYAML>=6.0.1
argparse>=1.4.0
# transformers>=4.39.3
djangorestframework
langchain-community
langchain-core
langchain
drf-yasg
django-filter
django-apscheduler
psycopg
psycopg2
html2text
python-docx
PyMuPDF
langchain-openai
pycryptodome
xlwt